Initial data to database
